requirements.txt
absl-py==1.3.0
-e git+https://github.com/aiortc/aioquic.git@239f99b8a3d4f5bc88cb280df765f35722cefe57#egg=aioquic
ale-py==0.7.4
anyio==3.6.1
asgiref==3.5.0
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
backports.zoneinfo==0.2.1
bcrypt==3.2.2
beautifulsoup4==4.10.0
bitarray==2.3.7
blinker==1.4
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
cloud-init==22.2
cloudpickle==1.6.0
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==37.0.2
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
distro-info===0.18ubuntu0.18.04.1
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
dnslib==0.9.21
dnspython==2.2.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.10
Flask==2.1.2
fonttools==4.29.1
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h11==0.13.0
httpbin==0.7.0
httplib2==0.9.2
hyperlink==17.3.1
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.10.1
incremental==16.10.1
influxdb==5.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.2
language-selector==0.1
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.3
msgpack==1.0.3
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
osm-im==9.0.0.post21+g024959e
osmclient==11.0.0rc1.post3+gd13d599
packaging==21.3
PAM==0.4.2
pandas==1.5.0
paramiko==2.11.0
pexpect==4.8.0
Pillow==9.0.1
pipreqs==0.4.11
protobuf==3.19.4
psutil==5.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyang==2.5.2
pyangbind==0.8.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pycrypto==2.6.1
pycurl==7.44.1
PyGObject==3.26.1
PyJWT==1.5.3
pylsqpack==0.3.16
pymongo==4.1.1
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyserial==3.4
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
python-dotenv==0.20.0
python-gvm==22.6.0
python-magic==0.4.25
pytz==2021.3
pyxdg==0.25
PyYAML==3.12
raven==6.10.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
rsa==4.8
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
sniffio==1.3.0
sos==4.3
soupsieve==2.3.1
sqlparse==0.4.2
ssh-import-id==5.7
starlette==0.20.4
systemd-python==234
texttable==1.6.4
timeout-decorator==0.5.0
tqdm==4.64.1
Twisted==17.9.0
typing_extensions==4.1.1
ubuntu-advantage-tools==27.8
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.8
verboselogs==1.7
websocket-client==0.59.0
Werkzeug==2.0.3
wsproto==1.2.0
xmltodict==0.13.0
yarg==0.1.9
zipp==3.7.0
zope.interface==4.3.2